secondary/requirements.txt
Flask==1.1.1
dependency-injector==4.3.2
protobuf
grpcio==1.33.2